Changed message lookup behavior so that translation(s) are tried first, then the default messages.xml.  This allows translations to only define those XML elements that actually contain translations, avoiding a lot of duplication.
Note that we could go further and allow the translations to override
only certain *parts* of the XML elements - for example, just the
ShortDescription and LongDescription elements of a BugPattern,
keeping the Details from the default messages.xml.  But, this would be
a lot more work.  The current scheme is reasonably simple and effective.


git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@2258 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3