[ 1651001 ] fancy-hist.xsl: stylesheet for historized analysis
This stylesheet is to be used with XML reports generated through mining features of findbugs.

It relies on patch "[1628150] new -withMessages option for mining tools".

The stylesheet is not complete yet (the 'browse by package' feature still need to be coded), but I want to get feedback from the community before going further.


Attached zip file contain as sample report as well as a sample build script that:
- runs findbugs
- add release info
- add last analysis to history XML report
- generate the out of the box change summary text file
- generates history-aware HTML report

git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@7530 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3