Null or empty output file doesn't set output file name
As suggested by   Blake Meike:

I would like to request a convenient change for the ant task.  I would like some value for the outputFile attribute that would make the task behave as if the attribute had not been specified.  That is, some value that would make the output go ant standard out.  I propose the empty string as the appropriate value.

This would be handy because it would allow the user of an ant build-file to control the output from findbugs, without editing the file (by changing the value of a property whose value is the name of the output file).  Currently, to change from output from standard out, to output to a file, it is necessary to edit the ant file.


git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@7559 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3