Improvements/documentation to BugAccumulator.
Experimental modification to FindDeadLocalStores to not report
a warning on a line if that line was observed to have a live
store on it at least once.  This is meant to eliminate some
false positives due to inlining/duplication of finally blocks.
However, it kills some real bugs when there are two stores
on the same line.  Currently, this modification is not enabled.

git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@8074 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3