Changed TypeQualifierApplications/TypeQualifierValue so that if a TypeQualifierValue is found without a when value, we make a note in order to remember to check it strictly.
Major refactoring of type qualifier dataflow analysis: sources and
sinks are computed once at the beginning of the analysis, keeping track
of all of the information needed to model them.
Eventually, we'll use this to more easily check when backwards
ALWAYS and NEVER flow values reach annotated sources.
A side benefit is that the logic of transferInstruction is
simplified and moved to TypeQualifierDataflowAnalysis (the superclass of the
forwards and backwards type qualifier dataflow analysis classes.)

git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@8529 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3