[ 1779315 ] Suggestion for new checks - File.delete
Y
We just had some problems with the following things:

- File.delete() returns a boolean if the delete operation was successful or
not. This is strange, as you would expect an IOException in such a case. Is
it possible to check if the return value of File.delete() is handled and
report if not as this indicates that the result of the delete is not
checked!

git-svn-id: https://findbugs.googlecode.com/svn/trunk/findbugs@8934 eae3c2d3-9b19-0410-a86e-396b6ccb6ab3