Use slashed class names and include patterns for notification in vbi
FindBugs prefers to use slashed class names instead of dotted ones, so the code was changed accordingly. It is not entirely clear to me whether that includes the descriptor and the semicolon. Because 'DescriptorFactory.createClassDescriptor' will not correctly identify classes if descriptor and semicolon are included in the string, I opted to not use them either.

The new pattern "VBC_MO_NOTIFY" is now detected when 'notify' or 'notifyAll' is called on a vbi.