Undeprecated findbugs annotations as using jsr305 is brain dead.
The defaultAnnotation bits are likely broken as they seem to rely on jsr
meta annotations to do their job.

there is NO SUCH THING AS JSR-305.
* The spec is non existent - not even a draft.
* There is only a snapshot of some code in a repository at some moment in
time.
* Should that jsr even become non dormant the annotations could change in a
non backwards compatible way
* What there is uses the javax namespace wich is reserved for official
specification.  As there is no specification this can not be official
* Shipping this with an application and an Oracle JDK is in direct
violation of the oracle licence agreement

This is considered your starter for 10 and a clue stick.