Fix outdated link in LICENSE-jsr305.txt
The old link does not exist anymore and now redirects to the
BSD-2-Clause page. The correct license for JSR-305 is
BSD-3-Clause. Following the old link requires extra clicks and
causes confusion, so this change updates the link to point
directly to BSD-3-Clause (aka "New BSD license") terms.